Begin to carve out platform/protocol API
