+++ tmk_core/protocol/vusb.mk
@@ -3,7 +3,7 @@ VUSB_DIR = protocol/vusb
 # Path to the V-USB library
 VUSB_PATH = $(LIB_PATH)/vusb
 
-SRC += $(VUSB_DIR)/main.c \
+SRC += $(VUSB_DIR)/protocol.c \
 	$(VUSB_DIR)/vusb.c \
 	$(VUSB_DIR)/usb_util.c \
 	$(VUSB_PATH)/usbdrv/usbdrv.c \
